requirements.txt
alembic==1.7.6
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
Flask==1.1.4
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==1.0.1
